refactor(sonar): fix Code Smells
- Default export names and file names should match
- Variables should not be shadowed
